server/requirements.txt
chroma-hnswlib==0.7.6
chromadb==0.5.23
groq==0.15.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.3.14
langchain-chroma==0.2.0
langchain-community==0.3.14
langchain-core==0.3.29
langchain-groq==0.2.3
langchain-ollama==0.2.2
langchain-text-splitters==0.3.5
pypdf==5.1.0
python-dotenv==1.0.1
python-json-logger==2.0.7